added checkstyle config
