+++ .editorconfig
@@ -9,3 +9,4 @@ end_of_line = lf
 charset = utf-8
 indent_style = tab
 insert_final_newline = true
+trim_trailing_whitespace = true
